requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
requests==2.31.0
redfish==3.0.1
python-dotenv